feat: move function to separate file
