+++ packages/superset-ui-dimension/src/getTextDimension.ts
@@ -1,4 +1,5 @@
 import { TextStyle, Dimension } from './types';
+import createSVGNode from './createSVGNode';
 
 const SVG_NS = 'http://www.w3.org/2000/svg';
 const STYLE_FIELDS: (keyof TextStyle)[] = [
@@ -18,18 +19,6 @@ export interface GetTextDimensionInput {
 }
 
 const DEFAULT_DIMENSION = { height: 20, width: 100 };
-
-export function createSVGNode(container: HTMLElement): SVGSVGElement {
-  const textNode = document.createElementNS(SVG_NS, 'text');
-  const svg = document.createElementNS(SVG_NS, 'svg');
-  svg.style.position = 'absolute'; // so it won't disrupt page layout
-  svg.style.opacity = '0'; // and not visible
-  svg.style.pointerEvents = 'none'; // and not capturing mouse events
-  svg.appendChild(textNode);
-  container.appendChild(svg);
-
-  return svg;
-}
 
 export default function getTextDimension(
   input: GetTextDimensionInput,
