+++ packages/superset-ui-dimension/src/index.ts
@@ -1,4 +1,5 @@
-export { default as getTextDimension, createSVGNode } from './getTextDimension';
+export { default as getTextDimension } from './getTextDimension';
+export { default as createSVGNode } from './createSVGNode';
 export { default as computeMaxFontSize } from './computeMaxFontSize';
 export { default as mergeMargin } from './mergeMargin';
 export { default as parseLength } from './parseLength';
